+++ src/index.ts
@@ -81,6 +81,7 @@
 export * from "./components/Form.js";
 export * from "./components/FormField.js";
 export * from "./components/ServerActionForm.js";
+export * from "./fields/AutocompleteField.js";
 export * from "./fields/CheckboxField.js";
 export * from "./fields/DateField.js";
 export * from "./fields/FileField.js";